Mapped out the first part of the notification builder. NotificationBuilder.organize_daily_email will be the function that is called by the scheduler.
NotificationBuilder.create_notify_list handles interaction with the Vote model, gathering what users need to be informed about which votes.

NotificationBuilder.key_value_crossover is required to correctly switch over the keys (being vote_ids) and the values (being an array of user_ids) to the opposite. As in, keys that are user_ids that have values of an array of vote_ids that the user needs to be informed about.