organize_daily_email acts as the controller for the rest of the NotificationsBuilder. It's logic is more or less defined as is.
The check_preferences method has been stubbed out. It's purpose is to take the notify list and compare each of the users in it to their notification preferences. If they don't want to recieve emails, those notifications will be stripped out of the notify_list

Added a email_test method, which successfully spams my account.

Added a defaultish email method.