Updated the .new_votes method on the Vote Model, to search on updated_at instead of created_at. This was changed to ensure that existing votes that are changed from one proposal to another are also captured as notifible activity.
Changed vote_spec in order to test the above.

Also added proposal existence check on the .find_users_in_tree method. This checks if there is an associated proposal for a vote, since there is no point running the logic of the method if there is not. The method will then return an empty array, which should have no impact on the methods that have a dependency on this method.

Tested the above in IRB